+++ R/AllClasses.R
@@ -6,9 +6,8 @@
 #' cell quality control metrics.
 #'
 #' @author Michael Steinbaugh, Rory Kirchner
+#' @note Updated 2019-07-30.
 #' @export
-
-## Updated 2019-07-24.
 setClass(
     Class = "bcbioSingleCell",
     contains = "SingleCellExperiment",