+++ R/data.R
@@ -1,5 +1,6 @@
 #' inDrops example data set
 #' @author Michael Steinbaugh
+#' @note Updated 2019-07-24.
 #' @examples
 #' data(indrops)
 #' indrops