+++ R/extract-methods.R
@@ -14,6 +14,8 @@
 #' @name extract
 #' @author Michael Steinbaugh
 #' @inherit base::Extract params references
+#' @note Updated 2019-07-24.
+#'
 #' @inheritParams basejump::params
 #'
 #' @return `bcbioSingleCell`.