+++ R/imports.R
@@ -14,11 +14,11 @@
 #' @importFrom acidplots acid_geom_bar acid_scale_y_continuous_nopad synesthesia
 #' @importFrom basejump camelCase cell2sample detectLanes emptyRanges
 #'   formalsList import interestingGroups interestingGroups<- makeDimnames
-#'   makeGRangesFromEnsembl makeGRangesFromGFF makeNames
-#'   makeSingleCellExperiment mapCellsToSamples markdownHeader markdownPlots
-#'   matchArgsToDoCall matchInterestingGroups metrics minimalSampleData
-#'   prepareTemplate printString relevel realpath sampleData separator
-#'   showSlotInfo
+#'   makeGRangesFromEnsembl makeGRangesFromGFF makeLabel makeNames
+#'   makeSingleCellExperiment makeTitle mapCellsToSamples markdownHeader
+#'   markdownPlots matchArgsToDoCall matchInterestingGroups metrics
+#'   minimalSampleData prepareTemplate printString relevel realpath sampleData
+#'   separator showSlotInfo
 #' @importFrom bcbioBase getBarcodeCutoffFromCommands getGTFFileFromYAML
 #'   getLevelFromCommands getSampleDataFromYAML getUMITypeFromCommands
 #'   projectDir readDataVersions readProgramVersions readSampleData runDate
@@ -37,7 +37,7 @@
 #'   isNonNegative isNumber isPositive isString isSubset validate
 #'   validateClasses
 #' @importFrom graphics hist
-#' @importFrom magrittr %>% set_names
+#' @importFrom magrittr %<>% %>% set_names
 #' @importFrom methods as as<- is new setClass show slot slot<- validObject
 #'   .hasSlot
 #' @importFrom acidplots acid_geom_abline acid_geom_label