+++ R/show-methods.R
@@ -2,6 +2,7 @@
 #' @name show
 #' @author Michael Steinbuagh
 #' @inherit methods::show
+#' @note Updated 2019-07-24.
 #' @examples
 #' data(indrops)
 #' show(indrops)