+++ tests/testthat/helper-globals.R
@@ -1,1 +1,3 @@
 data(indrops, envir = environment())
+
+hasInternet <- goalie::hasInternet