+++ NEWS.md
@@ -1,3 +1,17 @@
+## bcbioSingleCell 0.3.18 (2019-07-29)
+
+### Major changes
+
+- Updated R dependency to 3.6. This now requires Bioconductor 3.9+.
+- `updateObject`: Simplified method support, dropping `rowRanges` argument.
+  No longer attempting to coerce assays back to a simple list, which doesn't
+  work well on Bioconductor 3.10 Devel currently.
+
+### Minor changes
+
+- Updated basejump dependencies.
+- Improved code coverage over 90%.
+
 ## bcbioSingleCell 0.3.17 (2019-07-24)
 
 ### Minor changes